adding back the TFS protos because the query model from recommender_systems is now correctly exported to SavedModl format and deploys successfully to TFS. correcting dtypes and improving the graph-ranker model export and adding a validation report to it. The rust client has not been updated for the latest model exports, nor has the graph-ranker model been tested to deploy to TFS.

+++ src/main/python/movie_lens_ranker/SparseLocalSubgraphTransform.py
@@ -82,8 +82,8 @@ class SparseLocalSubgraphTransform(pgrain.MapTransform):
                 edges={"rating": edge_features},
                 senders=senders,
                 receivers=receivers,
-                n_node=np.array([total_nodes]),
-                n_edge=np.array([total_edges]),
+                n_node=np.array([total_nodes], dtype=np.int32),
+                n_edge=np.array([total_edges], dtype=np.int32),
                 globals=None
             ))
